🔥 remove google analytics
