+++ components/markets/MarketsTables/index.tsx
@@ -17,8 +17,6 @@ import MobileTabs from 'components/MobileTabs';
 import { TableHeader } from 'components/common/TableHeadCell';
 import useAccountData from 'hooks/useAccountData';
 import { useGlobalError } from 'contexts/GlobalErrorContext';
-import usePreviousValue from 'hooks/usePreviousValue';
-import useAnalytics from 'hooks/useAnalytics';
 import { MAX_UINT256, WEI_PER_ETHER } from 'utils/const';
 
 const { smOrLess, mdOrMore } = globals;
@@ -42,10 +40,6 @@ const MarketTables: FC = () => {
   const [fixedRows, setFixedRows] = useState<TableRow[]>([...defaultRows]);
   const [isLoading, setIsLoading] = useState<boolean>(true);
   const { setIndexerError } = useGlobalError();
-
-  const {
-    list: { viewItemListAdvance },
-  } = useAnalytics();
 
   const floatingHeaders: TableHeader<TableRow>[] = [
     {
@@ -189,19 +183,6 @@ const MarketTables: FC = () => {
   useEffect(() => {
     void defineRows();
   }, [defineRows]);
-
-  const previousFloatingRows = usePreviousValue(floatingRows);
-  const previousFixedRows = usePreviousValue(fixedRows);
-
-  useEffect(() => {
-    if (previousFloatingRows !== floatingRows && !isLoading) {
-      viewItemListAdvance(floatingRows, 'floating');
-    }
-
-    if (previousFixedRows !== fixedRows && !isLoading) {
-      viewItemListAdvance(fixedRows, 'fixed');
-    }
-  }, [floatingRows, fixedRows, previousFloatingRows, previousFixedRows, viewItemListAdvance, isLoading]);
 
   return (
     <>
