+++ contexts/GetEXAContext.tsx
@@ -34,7 +34,6 @@ import {
   Protocol,
   ActiveRoute,
   BridgeStatus,
-  BridgeInput,
 } from 'types/Bridge';
 import useSocketAssets from 'hooks/useSocketAssets';
 import handleOperationError from 'utils/handleOperationError';
@@ -63,7 +62,6 @@ import waitForTransaction from 'utils/waitForTransaction';
 import dayjs from 'dayjs';
 import { splitSignature } from '@ethersproject/bytes';
 import useDelayedEffect from 'hooks/useDelayedEffect';
-import useAnalytics from '../hooks/useAnalytics';
 
 const DESTINATION_CHAIN = optimism.id;
 
@@ -141,8 +139,6 @@ export const GetEXAProvider: FC<PropsWithChildren> = ({ children }) => {
     () => assets.filter(({ chainId }) => chainId === chain?.chainId),
     [assets, chain?.chainId],
   );
-  const { transaction } = useAnalytics();
-
   const exaPrice = useEXAPrice();
   const isBridge = chain?.chainId !== appChain.id;
 
@@ -398,18 +394,6 @@ export const GetEXAProvider: FC<PropsWithChildren> = ({ children }) => {
         : 0n
       : undefined;
 
-  const input: BridgeInput = useMemo(
-    () => ({
-      destinationAmount: qtyOut?.toString() || '',
-      destinationToken: 'EXA',
-      sourceAmount: qtyIn,
-      sourceToken: asset?.symbol || '',
-      destinationChainId: DESTINATION_CHAIN,
-      sourceChainId: chain?.chainId,
-    }),
-    [asset?.symbol, chain?.chainId, qtyIn, qtyOut],
-  );
-
   const confirmBridge = useCallback(async () => {
     if (txStep !== TXStep.CONFIRM || !walletClient || !route) return;
 
@@ -426,20 +410,15 @@ export const GetEXAProvider: FC<PropsWithChildren> = ({ children }) => {
       const { status, transactionHash } = await waitForTransaction({ hash: txHash_ });
       setTX({ status: status ? 'success' : 'error', hash: transactionHash });
       setScreen(Screen.TX_STATUS);
-      if (status === 'success') {
-        transaction.purchase('getEXA', input);
-      }
     } catch (err) {
       setTXError({ status: true, message: handleOperationError(err) });
       setTXStep(TXStep.CONFIRM);
-      transaction.removeFromCart('getEXA', input);
     }
-  }, [destinationCallData, input, route, transaction, txStep, walletClient]);
+  }, [destinationCallData, route, txStep, walletClient]);
 
   const socketSubmit = useCallback(async () => {
     const minEXA = 0n;
     const keepETH = 0n;
-    transaction.beginCheckout('getEXA', input);
     if (isBridge) return confirmBridge();
 
     if (!walletAddress || !route || !swapper || !erc20?.address || !opts || !asset) return;
@@ -488,32 +467,14 @@ export const GetEXAProvider: FC<PropsWithChildren> = ({ children }) => {
         setScreen(Screen.TX_STATUS);
         setTX({ status: 'processing', hash });
         const { status, transactionHash } = await waitForTransaction({ hash });
-        if (status === 'success') {
-          transaction.purchase('getEXA', input);
-        }
         setTX({ status: status ? 'success' : 'error', hash: transactionHash });
       }
     } catch (err) {
       setTXError({ status: true, message: handleOperationError(err) });
-      transaction.removeFromCart('getEXA', input);
     } finally {
       setTXStep(undefined);
     }
-  }, [
-    transaction,
-    input,
-    isBridge,
-    confirmBridge,
-    walletAddress,
-    route,
-    swapper,
-    erc20,
-    opts,
-    asset,
-    isMultiSig,
-    qtyIn,
-    sign,
-  ]);
+  }, [isBridge, confirmBridge, walletAddress, route, swapper, erc20, opts, asset, isMultiSig, qtyIn, sign]);
 
   const submit = useCallback(async () => {
     if (!walletClient || !walletAddress || !swapper) return;
