+++ package.json
@@ -48,7 +48,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-draggable": "^4.4.5",
-    "react-ga4": "^2.1.0",
     "react-i18next": "^13.0.3",
     "react-number-format": "^5.2.2",
     "recharts": "^2.7.2",
