+++ pages/_app.tsx
@@ -16,7 +16,7 @@ import Footer from 'components/Footer';
 import Navbar from 'components/Navbar';
 import { globals } from 'styles/theme';
 import { GlobalErrorProvider } from 'contexts/GlobalErrorContext';
-import { useInitGA, usePageTracking } from 'hooks/useAnalytics';
+import { usePageTracking } from 'hooks/usePageTracking';
 
 import { ModalContextProvider } from 'contexts/ModalContext';
 import OperationsModal from 'components/OperationsModal';
@@ -45,7 +45,6 @@ const Modals = () => (
 );
 
 export default function App({ Component, pageProps, router }: AppProps) {
-  useInitGA();
   usePageTracking();
 
   return (
