+++ pages/bridge.tsx
@@ -1,10 +1,8 @@
 import React, { memo } from 'react';
 
 import BridgeContent from 'components/BridgeContent';
-import { usePageView } from 'hooks/useAnalytics';
 
 const Bridge = () => {
-  usePageView('/bridge', 'Bridge');
   return <BridgeContent />;
 };
