+++ pages/get-exa.tsx
@@ -3,13 +3,10 @@ import { NextPage } from 'next';
 import { Box, Typography } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
 import { GetEXAProvider } from '../contexts/GetEXAContext';
-import { usePageView } from '../hooks/useAnalytics';
 import GetEXA from '../components/GetEXA';
 import Link from 'next/link';
 
 const GetExaPage: NextPage = () => {
-  usePageView('/get-exa', 'Get EXA');
-
   const { t } = useTranslation();
 
   return (
