+++ pages/index.tsx
@@ -7,11 +7,9 @@ import MarketTables from 'components/markets/MarketsTables';
 import MarketsBasic from 'components/markets/MarketsBasic';
 import BackgroundCircle from 'components/BackgroundCircle';
 
-import { usePageView } from 'hooks/useAnalytics';
 import { useCustomTheme } from 'contexts/ThemeContext';
 
 const Markets: NextPage = () => {
-  usePageView('/', 'Markets');
   const { view } = useCustomTheme();
 
   if (!view) return null;
