+++ pages/revoke.tsx
@@ -1,9 +1,7 @@
 import Allowances from 'components/Allowances';
-import { usePageView } from 'hooks/useAnalytics';
 import React, { memo } from 'react';
 
 const Revoke = () => {
-  usePageView('/revoke', 'Manage Allowances');
   return <Allowances />;
 };
