+++ pages/security/protocol.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import type { NextPage } from 'next';
 
-import { usePageView } from 'hooks/useAnalytics';
 import { useTranslation } from 'react-i18next';
 import { Box, Button, Skeleton, Typography } from '@mui/material';
 import ArrowForwardRoundedIcon from '@mui/icons-material/ArrowForwardRounded';
@@ -15,7 +14,6 @@ import { useWeb3 } from 'hooks/useWeb3';
 import { mainnet, goerli, optimism } from 'viem/chains';
 
 const Security: NextPage = () => {
-  usePageView('/security', 'Security');
   const { t } = useTranslation();
   const { query } = useRouter();
   const [contractsData, setContractsData] = useState<ContractInfoType[]>([]);
