--- types/Vest.ts
@@ -1,8 +1,0 @@
-export type VestInput = {
-  sourceChainId?: number;
-  sourceToken: string;
-  destinationChainId?: number;
-  destinationToken: string;
-  sourceAmount: string;
-  destinationAmount: string;
-};
